perf: optimize code
